+++ src/index.html
@@ -31,6 +31,5 @@
     </form>
 
     <script type="text/javascript" src="scripts/main.js"></script>
-    <script type="text/javascript" src="scripts/callPromise.js"></script>
   </body>
 </html>
