+++ src/scripts/main.js
@@ -1,24 +1,35 @@
 'use strict';
 
+// Definicja funkcji waitFor
 function waitFor(element, eventName) {
   return new Promise((resolve, reject) => {
     element.addEventListener(`${eventName}`, () => {
-      resolve(`It was ${eventName} on the element:
-        ${element.nodeName}, id: ${element.id}.`);
+      resolve(`It was ${eventName} on the element: ${element.nodeName}, id: ${element.id}.`);
     });
   });
 }
 
-const printMessage = (message) => {
+// Definicja funkcji printMessage
+function printMessage(message) {
   document.body.insertAdjacentHTML(
     'beforeend',
-    `
-    <div class="message">${message}</div>
-  `,
+    `<div class="message">${message}</div>`
   );
-};
+}
 
-module.exports = {
-  waitFor,
-  printMessage,
-};
+// Pobranie referencji do elementów formularza
+const loginField = document.getElementById('login');
+const passwordField = document.getElementById('password');
+const button = document.getElementById('submit');
+
+// Obsługa zdarzeń dla pól formularza i przycisku
+waitFor(loginField, 'click').then(printMessage);
+waitFor(passwordField, 'click').then(printMessage);
+waitFor(button, 'click').then(printMessage);
+
+waitFor(loginField, 'input').then(printMessage);
+waitFor(passwordField, 'input').then(printMessage);
+
+waitFor(loginField, 'blur').then(printMessage);
+waitFor(passwordField, 'blur').then(printMessage);
+waitFor(button, 'blur').then(printMessage);
